+++ src/lib.rs
@@ -4,4 +4,4 @@ pub mod filter;
 pub mod types;
 pub mod spatial;
 pub mod testing_util;
-pub mod worker_utils;
+pub mod worker_util;
